+++ lib/commands.ml
@@ -9,8 +9,8 @@ let printfn = Util.printfn
 module Init = struct
   let init ?use_sudo ?(force = false) () =
     let base_dir = Lazy.force !Config.base_dir in
-    if Sys.is_directory base_dir && not force then
-      die "E: Passage init failed, %s already exists (add --force to overwrite)" base_dir;
+    let has_config = try Sys.is_directory base_dir with _ -> false in
+    if has_config && not force then die "E: Passage init failed, %s already exists (add --force to overwrite)" base_dir;
     try
       (* create private and pub key, ask for user's name *)
       let () =
@@ -35,7 +35,7 @@ What should be the name used for your recipient identity?|}
       in
       let user_name =
         match In_channel.input_line stdin with
-        | None -> die "E: EOF while reading usern ame."
+        | None -> die "E: EOF while reading user name."
         | Some line ->
           let line = String.trim line in
           let buf = Buffer.create String.(length line) in
@@ -47,7 +47,7 @@ What should be the name used for your recipient identity?|}
             line;
           Buffer.contents buf
       in
-      if Sys.is_directory base_dir && force then FileUtil.rm ~recurse:true [ base_dir ];
+      if has_config && force then FileUtil.rm ~recurse:true [ base_dir ];
       let () = Shell.age_generate_identity_key_root_group_exn ?use_sudo user_name in
       verbose_eprintlf "I: Passage setup completed.\n"
     with exn ->